Move debezium builder into standalone function common project
 - Move the DebeziumEngineBuilderAutoConfiguration and META-INF configurations to new function project: debezium-autoconfigure.
 - Introduce the EmbeddedEngineExecutorService helper class for help DebezuimEngine implementations.
 - Move the relevat tests from debezium-supplier to the debeizum-autoconfigure.
 - Add detailer README documentation.
 - Adjust the debeizum-supplier and debezium-source configurations.
 - Clean the related tests.

 Resolves #462
