+++ functions/common/debezium-auto-configuration/src/main/java/org/springframework/cloud/fn/common/debezium/DebeziumProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.fn.supplier.debezium;
+package org.springframework.cloud.fn.common.debezium;
 
 import java.util.HashMap;
 import java.util.Map;
