+++ functions/common/debezium-auto-configuration/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+org.springframework.cloud.fn.common.debezium.DebeziumEngineBuilderAutoConfiguration
