--- functions/supplier/debezium-supplier/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-org.springframework.cloud.fn.supplier.debezium.DebeziumEngineBuilderAutoConfiguration
